Provide public test utils to create PersonDetection or POISnapshot models
