+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@advertima/io",
-  "version": "0.4.0",
+  "version": "0.4.1",
   "description": "IO utilities to connect to a PoI",
   "main": "lib/io.cjs.js",
   "module": "lib/io.esm.js",
@@ -84,7 +84,10 @@
   },
   "nyc": {
     "include": "**/*.ts",
-    "exclude": "**/*.spec.ts",
+    "exclude": [
+      "**/*.spec.ts",
+      "src/poi/test-utils"
+    ],
     "extension": [
       ".ts"
     ]
