+++ src/index.ts
@@ -6,6 +6,7 @@ export * from './messages/MessageFactory';
 export * from './messages/person-detection/PersonDetectionMessage';
 export * from './messages/content/ContentMessage';
 export * from './messages/persons-alive/PersonsAliveMessage';
+export * from './messages/skeleton/SkeletonMessage';
 export * from './messages/unknown/UnknownMessage';
 export * from './incoming-stream/IncomingStream';
 export * from './model/person-detection/PersonDetection';
@@ -16,3 +17,4 @@ export * from './event/EventMonitor';
 export * from './constants/Constants';
 export * from './io/IO';
 export * from './types';
+export { PersonOptions, PersonDetectionGenerator, POISnapshotGenerator } from './poi/test-utils';
