+++ src/poi/test-utils/index.ts
@@ -1,4 +1,7 @@
-export * from './messages/PersonDetectionMessageGenerate';
-export * from './messages/SkeletonMessageGenerate';
-export * from './messages/PersonsAliveMessageGenerate';
-export * from './messages/ContentMessageGenerate';
+export * from './messages/PersonDetectionMessageGenerator';
+export * from './messages/SkeletonMessageGenerator';
+export * from './messages/PersonsAliveMessageGenerator';
+export * from './messages/ContentMessageGenerator';
+export * from './poi/POISnapshotGenerator';
+export * from './model/PersonDetectionGenerator';
+export * from './common';
