--- src/poi/test-utils/messages/PersonsAliveMessageGenerate.ts
@@ -1,22 +1,0 @@
-import { PersonsAliveMessage } from '../../../messages/persons-alive/PersonsAliveMessage';
-import { MessageFactory } from '../../../messages/MessageFactory';
-import { RPCResponseSubject } from '../../../constants/Constants';
-
-/**
- * Generates a PersonsAliveMessage instance
- * @param {string[]} ids list of person id
- * @param {number} timestamp optional timestamp
- * @return {PersonsAliveMessage} message
- */
-export function generatePersonsAliveMessage(
-  ids: string[] = [],
-  timestamp: number = Date.now()
-): PersonsAliveMessage {
-  return MessageFactory.parse({
-    subject: RPCResponseSubject.PersonsAlive,
-    data: {
-      person_ids: ids,
-      local_timestamp: timestamp,
-    },
-  }) as PersonsAliveMessage;
-}
